backend/app/deps/hashed_password.py
import jwt

from app.db.config import algorithm, secret_key
from jwt.exceptions import InvalidTokenError
from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
from pwdlib import PasswordHash

ALGORITHM = algorithm
SECRET_KEY = secret_key
ACCESS_TOKEN_EXPIRE_MINUTES = 30


password_hash = PasswordHash.recommended()

oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/token")


def verify_password(plain_password: str, hashed_password: str) -> bool:
    return password_hash.verify(plain_password, hashed_password)

def get_password_hashed(password: str) -> str:
    return password_hash.hash(password)